\name{DipIndex}
\alias{DipIndex}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Diploidization Index
}
\description{
Indexes and compares recovered alleles for diploids and tetraploids to determine level of diploidization that has occured in tetraploids at each locus for each individual}
\usage{
DipIndex()
}

\author{
William Booker}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
DipIndex
}