\name{MakeGraphs}
\alias{MakeGraphs}
\alias{LocusPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Make Allele Distance Graphs}
\description{
graphs allelic distances for the two least smallest distances for each individual locus. points on those lines are individuals, green individuals are unknown, blue are tetraploids, and red are diploids. \cr\cr

Stored as a pdf in the PloidyRFiles folder}
\usage{
MakeGraphs()
}
\author{
William Booker}

\keyword{ PloidyPal }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
