\name{ReadSumFromFile}
\alias{ReadSumFromFile}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read Summary Info from File}
\description{
Takes in info as .csv files that was generated in previous runs of CheckPloidy() to use in other functions}
\usage{
ReadSumFromFile(rootPath)
}
%- maybe also 'usage' for other objects documented here.
\author{
William Booker
}


